requirements.txt
appdirs==1.4.3
asana==0.9.2
asgiref==3.2.3
astroid==2.3.3
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
distlib==0.3.0
dj-database-url==0.5.0
Django==3.0.3
django-filter==2.2.0
django-heroku==0.3.1
djangorestframework==3.11.0
filelock==3.0.12
gunicorn==20.0.4
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.3
Markdown==3.2.1
mccabe==0.6.1
oauthlib==3.1.0
pipenv==2018.11.26
psycopg2==2.8.4
pylint==2.4.4
pylint-django==2.0.13
pylint-plugin-utils==0.6
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
six==1.14.0
sqlparse==0.3.0
urllib3==1.25.8
virtualenv==20.0.2
virtualenv-clone==0.5.3
whitenoise==5.0.1
wrapt==1.11.2
